Move tangoFormatter to a separated module
Move tangoFormatter from taurus.core.tango.util.tango_taurus
to taurus.core.tango.util.formatter.

It can be imported like:
`from taurus.core.tango.util import tangoFormatter`
