+++ lib/taurus/core/tango/util/__init__.py
@@ -26,3 +26,5 @@
 """The sardana package. It contains specific part of sardana"""
 
 __docformat__ = 'restructuredtext'
+
+from formatter import *
